+++ CHANGES.md
@@ -24,6 +24,7 @@
 * change get address field
 * rename traverse to traverse_needed (register)
 * add direct_proxy to cmdcom
+* do_request can add pw to returned dict
 
 ## bugfixes
 * better verification serveranswers
@@ -34,6 +35,7 @@
 * fix blocked pw requests by moving composed requests
 * fix cleaning certreferences
 * fix traversal (untested)
+
 
 # 0.6.2
 ## bugfixes